backend/requirements.txt
django>=5.1,<6.0
python-dotenv>=1.1,<2.0
